update velocity controller to use velocity float (combined motor rpms) to soon deprecate WSS msg
