+++ src/control/velocity_controller/package.xml
@@ -8,6 +8,9 @@
     <license>MIT</license>
 
     <!-- Messages -->
+    <depend>ackermann_msgs</depend>
+    <depend>nav_msgs</depend>
+    <depend>rcl_interfaces</depend>
     <depend>driverless_msgs</depend>
 
     <buildtool_depend>ament_cmake</buildtool_depend>
